refactor: remove useless tooltip-helper abstract class
