+++ packages/vchart/src/component/interface/index.ts
@@ -1,4 +1,3 @@
 export * from './common';
 export * from './type';
-export * from './tooltip-helper';
 export * from './theme';
