--- packages/vchart/src/component/interface/tooltip-helper.ts
@@ -1,6 +1,0 @@
-import type { ITooltipHelper } from '../../model/tooltip-helper';
-import type { IComponent } from './common';
-
-export interface IComponentTooltipHelper extends ITooltipHelper {
-  component: IComponent;
-}
