+++ packages/vchart/src/model/interface.ts
@@ -20,6 +20,8 @@ import type { ILayoutItem, ILayoutItemSpec } from '../layout/interface';
 import type { ILayoutPoint, ILayoutRect } from '../typings/layout';
 import type { ComponentTypeEnum } from '../component/interface';
 import type { SeriesTypeEnum } from '../series';
+import type { ITooltipSpec } from '../component/tooltip/interface';
+import type { TooltipActiveType } from '../typings';
 
 // TODO:
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
@@ -238,4 +240,25 @@ export interface IBaseModelSpecTransformer {
     chartSpec: any,
     chartSpecInfo?: IChartSpecInfo
   ) => IBaseModelSpecTransformerResult<any, any>;
+}
+
+export interface ITooltipHelper {
+  /** tooltip对应spec */
+  spec: ITooltipSpec | undefined;
+
+  /** 实际生效的tooltip activeType */
+  activeType: TooltipActiveType[];
+
+  /** 可以响应mark tooltip或者dimension tooltip的对象 */
+  activeTriggerSet: {
+    mark?: Set<IMark>;
+    group?: Set<IMark>;
+  };
+  /** 不响应tooltip且不会影响已有tooltip的对象 */
+  ignoreTriggerSet: {
+    mark?: Set<IMark>;
+  };
+
+  /** 更新spec */
+  updateTooltipSpec: () => void;
 }
